rename package from analytics to segment
