+++ segment/client.py
@@ -6,9 +6,9 @@ import numbers
 from dateutil.tz import tzutc
 from six import string_types
 
-from analytics.utils import guess_timezone, clean
-from analytics.consumer import Consumer
-from analytics.version import VERSION
+from segment.utils import guess_timezone, clean
+from segment.consumer import Consumer
+from segment.version import VERSION
 
 try:
     import queue
