+++ segment/test/__init__.py
@@ -5,7 +5,7 @@ import sys
 
 def all_names():
     for _, modname, _ in pkgutil.iter_modules(__path__):
-        yield 'analytics.test.' + modname
+        yield 'segment.test.' + modname
 
 def all():
     logging.basicConfig(stream=sys.stderr)
