+++ segment/test/client.py
@@ -3,8 +3,8 @@ import unittest
 import time
 import six
 
-from analytics.version import VERSION
-from analytics.client import Client
+from segment.version import VERSION
+from segment.client import Client
 
 
 class TestClient(unittest.TestCase):
