+++ segment/utils.py
